+++ changelog.md
@@ -1,5 +1,9 @@
 # TreeQuest Changelog:
 
+### v0.1.2 (TELL ME! [PROLOGUE])
+ - Implements "notification" system
+ - All item pickups now give an "obtained item" notification
+
 ### v0.1.1 (Deathproof [PROLOGUE])
  - No longer forces you to return to Maintenance when oxygen runs out
  - Maintenance layer now highlights when oxygen runs out